Reworked TCP packet parsing
git-svn-id: https://svn.ntop.org/svn/ntop/trunk/PF_RING@4305 a02cd0c1-8e76-42e1-a119-56f9641475e2

+++ userland/examples/ptwin.c
@@ -115,7 +115,7 @@ void print_stats() {
       diff = pcapStat.ps_recv-lastPkts;
       fprintf(stderr, "=========================\n"
 	      "Actual Stats: %llu pkts [%.1f ms][%.1f pkt/sec]\n",
-	      diff, deltaSec*1000, ((double)diff/(double)(deltaSec)));
+	      (long long unsigned int)diff, deltaSec*1000, ((double)diff/(double)(deltaSec)));
       lastPkts = pcapStat.ps_recv;
     }
